refactor(schema): single zod source of truth for docs frontmatter (S7 #2016)
Collapses the 3 hand-mirrored docs-frontmatter shapes into one exported
schema builder in src/config/docs-schema.ts. DocsData (z.infer) replaces
the manual field lists in pages/_data.ts (ZfbDocsData alias) and
src/types/docs-entry.ts (DocsEntry.data). The generator now emits
`import { buildDocsSchema } from "./src/config/docs-schema"` and
`const docsSchema = buildDocsSchema()` instead of an inline field list,
so tag governance is encapsulated in the schema builder without extra
per-feature logic in zfb-config-gen.ts. Template files mirror host exactly.
Generator tests updated to assert the new import/delegation pattern.
